Wrap messages in a newtype bundle
